requirements.txt
numpy==1.19.5
wordcloud==1.8.1
tqdm==4.62.2
jieba==0.42.1
pandas==1.1.5
matplotlib==3.0.1
requests==2.26.0
pyquery==1.4.3
Pillow==8.3.2
